requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dateutil==2.8.2
python-frontmatter==1.1.0